+++ src/collective/solr/search.py
@@ -46,8 +46,10 @@ class Search(object):
             # http://wiki.apache.org/solr/CommonQueryParameters#rows
             if parameters['rows'] == 0:
                 parameters['rows'] = 10000000
-            logger.info('falling back to "max_results" (%d) without a "rows" '
-                'parameter: %r (%r)', config.max_results, query, parameters)
+            logger.debug(
+                'falling back to "max_results" (%d) without a "rows" '
+                'parameter: %r (%r)', config.max_results, query, parameters
+            )
         if getattr(config, 'highlight_fields', None):
             if parameters.get('hl', 'false') == 'true' and not 'hl.fl' in parameters:
                 parameters['hl'] = 'true'
@@ -76,9 +78,13 @@ class Search(object):
         elapsed = (time() - start) * 1000
         slow = config.slow_query_threshold
         if slow and elapsed >= slow:
-            logger.info('slow query: %d/%d ms for %r (%r)',
-                results.responseHeader['QTime'], elapsed, query, parameters)
-        logger.debug('highlighting info: %s' % getattr(results, 'highlighting', {}))
+            logger.info(
+                'slow query: %d/%d ms for %r (%r)',
+                results.responseHeader['QTime'], elapsed, query, parameters
+            )
+        logger.debug(
+            'highlighting info: %s' % getattr(results, 'highlighting', {})
+        )
         return results
 
     __call__ = search
@@ -93,15 +99,22 @@ class Search(object):
         for name, value in sorted(args.items()):
             field = schema.get(name or defaultSearchField, None)
             if field is None or not field.indexed:
-                logger.info('dropping unknown search attribute "%s" '
-                    ' (%r) for query: %r', name, value, args)
+                logger.info(
+                    'dropping unknown search attribute "%s" '
+                    ' (%r) for query: %r', name, value, args
+                )
                 continue
             if isinstance(value, bool):
                 value = str(value).lower()
             elif not value:     # solr doesn't like empty fields (+foo:"")
                 if not name:
                     continue
-                logger.info('empty search term form "%s:%s", aborting buildQuery' % (name,value))
+                logger.info(
+                    'empty search term form "%s:%s", aborting buildQuery' % (
+                        name,
+                        value
+                    )
+                )
                 return {}
             elif field.class_ == 'solr.BoolField':
                 if not isinstance(value, (tuple, list)):
@@ -144,8 +157,9 @@ class Search(object):
                 if not value:   # don't search for empty strings, even quoted
                     continue
             else:
-                logger.info('skipping unsupported value "%r" (%s)',
-                    value, name)
+                logger.info(
+                    'skipping unsupported value "%r" (%s)', value, name
+                )
                 continue
             if name is None:
                 if value and value[0] not in '+-':